# feat(tests): faster smokes via process-compose