--- a/justfile
+++ b/justfile
@@ -1,0 +1,4 @@
+smoke:
+    # resetting network state
+    cargo run --release --bin pd -- testnet unsafe-reset-all || true
+    ./deployments/scripts/smoke-test.sh