### Overview
A while back we ditched using containers for the smoke tests, mostly because the caching on the container-build story was atrocious, so test re-runs took a really long time. And frankly, container ergonomics on dev workstations, particularly macOS, are not awesome. Instead, let's assume the dev env can run processes for cargo, pd, and cometbft. If so, that's all we need to wire up our integration testing. Enter [process-compose](https://github.com/F1bonacc1/process-compose).

### Implementation
The new smoke test setup ditches the bash script and delegates to process-compose for orchestrating processes:

0. pre-building binaries and test artifacts
1. testnet config generation
2. running pd and cometbft
3. running `cargo test` for each binary's integration tests
4. reporting output

The wrapper script uses tty detection, so that running the smoke tests locally on your workstation displays a pleasant TUI:

![smoke-tui-1](https://github.com/penumbra-zone/penumbra/assets/657862/e27f9686-c33a-4e13-9b93-c2b3d26b8b71)


Logging is currently a bit too aggregated, would be nice to have discrete info about what failed.

When running in CI, however, log output will be displayed:

![smoke-test-log-output-ci](https://github.com/penumbra-zone/penumbra/assets/657862/9c8600d4-98ff-4957-8a81-71e3708cd21d)

### Benchmarks
tl;dr running the smoke tests in CI now takes about ~10m, down from ~20m [as in this recent run](https://github.com/penumbra-zone/penumbra/actions/runs/8949706068/job/24584307282?pr=4316). With a warm cache locally, run times are as low as ~3m30s, although the CI caching setup isn't as complete as local context. 

Testing the old script against the new invocation locally via hyperfine, results were:

<details>
<summary>
hyperfine results
</summary>

```
hyperfine --runs 3 --prepare 'cargo run -q --bin pd -- testnet unsafe-reset-all || true' -n new-smoke ./deployments/scripts/smoke-test.sh -n legacy-s
moke ./deployments/scripts/smoke-test-legacy.sh
Benchmark 1: new-smoke
  Time (mean ± σ):     191.469 s ±  1.009 s    [User: 421.365 s, System: 25.992 s]
  Range (min … max):   190.442 s … 192.459 s    3 runs

Benchmark 2: legacy-smoke
  Time (mean ± σ):     459.222 s ± 11.546 s    [User: 422.412 s, System: 22.178 s]
  Range (min … max):   445.889 s … 465.935 s    3 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It
 might help to use the '--warmup' or '--prepare' options.

Summary
  'new-smoke' ran
    2.40 ± 0.06 times faster than 'legacy-smoke'
```
</details>

Actual run times on this PR:

1. [10m22s](https://github.com/penumbra-zone/penumbra/actions/runs/8971907894/job/24638693639?pr=4324)
2. [8m18s](https://github.com/penumbra-zone/penumbra/actions/runs/8972056091/job/24639182542?pr=4324)
3. [10m50s](https://github.com/penumbra-zone/penumbra/actions/runs/8972313596/job/24640025148?pr=4324)