@conorsch 2024-05-06
> With a warm cache locally, run times are as low as ~3m30s, although the CI caching setup isn't as complete as local context.

Rerunning locally, I get super fast builds. I'm not going to drill into the caching story in this PR, because that's a bit of a rabbit hole, and a 2x speed-up is a big enough win that I want to merge this. It's worth exploring caching improvements because we'll likely have a lot of caching needs for #4323, too.