@conorsch 2024-05-06
I'm just going to say explicitly that this PR is dedicated to @TalDerei, whose brave sacrifices during #4081 will be remembered. I tried to build the kind of setup that would avoid a lot of the friction you experienced when wrestling with the smoke tests locally.