@conorsch 2024-05-06
> process-compose a more friendly container orchestration tool than say standard docker-compose or kubernetes? 

Simply because running containers locally has been a bit of a bear in the past. A prior version of the smoke test setup used docker-compose, and it was not pleasant to work with. We've since updated the container build story enough that it's worth revisiting, particularly in the context of upgrade testing, but this is a "good enough" solution for an immediate win, and leaves the door open to more robust devenvs via nix/devbox in the future.

> Also what's the scope of the information stored in the local hot cache?

The `target/` directory for rust build artifacts. On multiple CI runs, I still see some of the test artifacts being re-compiled, which indicates the caching is not perfect. It's still pretty darn good, though.